test(agentic-ai): split provider tests into per-provider classes
Move each provider's builder-level tests out of the monolithic
ChatModelFactoryTest into its own test class next to the provider,
and reduce ChatModelFactoryTest to its real responsibility: verifying
that the factory delegates to the registry-selected provider. Shared
helpers (timeout streams, default config binding, ResultCaptor) live
in a new ChatModelProviderTestSupport utility.
